+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## 7.6.0
-- TODO
+
+### Features
+- [#1507](https://github.com/okta/okta-auth-js/pull/1507) add: new method `getOrRenewAccessToken`
 
 ## 7.5.1